Move contact to CRM directory
